[IDB IGNORE] [MDB IGNORE] Baymerge 2025

+++ code/modules/detectivework/tools/swabs.dm
@@ -4,20 +4,18 @@
 	icon = 'icons/obj/forensics.dmi'
 	icon_state = "swab"
 
-/obj/item/swabber/attack()
-	return 0
-
 // This is pretty nasty but is a damn sight easier than trying to make swabs a stack item.
-/obj/item/swabber/afterattack(atom/A, mob/user, proximity, params)
-	if(proximity)
-		var/obj/item/forensics/swab/swab = new(user)
-		var/resolved = swab.resolve_attackby(A, user, params)
-		if(!resolved && A && !QDELETED(A))
-			swab.afterattack(A, user, TRUE, params)
-		if(swab.is_used())
-			swab.dropInto(user.loc)
-		else
-			qdel(swab)
+/obj/item/swabber/use_before(atom/target, mob/living/user, click_parameters)
+	if (user.a_intent == I_HELP) // Prevents putting sample kits in bags, on racks/tables, etc when trying to take samples
+		return ..()
+
+	var/obj/item/forensics/swab/swab = new(user)
+	swab.resolve_attackby(target, user, click_parameters)
+	if(swab.is_used())
+		swab.dropInto(user.loc)
+	else
+		qdel(swab)
+	return TRUE
 
 /obj/item/forensics/swab
 	name = "swab"
@@ -31,71 +29,64 @@
 /obj/item/forensics/swab/proc/is_used()
 	return used
 
-/obj/item/forensics/swab/attack(mob/living/M, mob/user)
+/obj/item/forensics/swab/use_before(mob/living/M, mob/user)
+	. = FALSE
+	if (!ishuman(M))
+		return FALSE
 
-	if(!ishuman(M))
-		return ..()
-
-	if(is_used())
-		return
+	if (is_used())
+		to_chat(user, SPAN_WARNING("This swab has already been used."))
+		return TRUE
 
 	var/mob/living/carbon/human/H = M
 	var/sample_type
 
-	if(H.wear_mask)
-		to_chat(user, SPAN_WARNING("\The [H] is wearing a mask."))
-		return
-
-	if(!H.dna || !H.dna.unique_enzymes)
+	if (!H.dna || !H.dna.unique_enzymes)
 		to_chat(user, SPAN_WARNING("They don't seem to have DNA!"))
-		return
+		return TRUE
 
-	if(user != H && (H.a_intent != I_HELP && !H.lying && !H.incapacitated(INCAPACITATION_DEFAULT)))
+	if (user != H && (H.a_intent != I_HELP && !H.lying && !H.incapacitated(INCAPACITATION_DEFAULT)))
 		user.visible_message(SPAN_DANGER("\The [user] tries to take a swab sample from \the [H], but they move away."))
-		return
+		return TRUE
 
-	if(user.zone_sel.selecting == BP_MOUTH)
-		if(!H.organs_by_name[BP_HEAD])
+	if (user.zone_sel.selecting == BP_MOUTH)
+		if (!H.organs_by_name[BP_HEAD])
 			to_chat(user, SPAN_WARNING("They don't have a head."))
-			return
-		if(!H.check_has_mouth())
+			return TRUE
+		if (!H.check_has_mouth())
 			to_chat(user, SPAN_WARNING("They don't have a mouth."))
-			return
+			return TRUE
+		if (H.wear_mask)
+			to_chat(user, SPAN_WARNING("\The [H] is wearing a mask."))
+			return TRUE
 		user.visible_message("[user] swabs \the [H]'s mouth for a saliva sample.")
 		dna = list(H.dna.unique_enzymes)
 		sample_type = "DNA"
 
 	else
 		var/zone = user.zone_sel.selecting
-		if(!H.has_organ(zone))
+		if (!H.has_organ(zone))
 			to_chat(user, SPAN_WARNING("They don't have that part!"))
-			return
+			return TRUE
 		var/obj/item/organ/external/O = H.get_organ(zone)
-		if(!O.gunshot_residue)
-			return
+		if (!O.gunshot_residue)
+			to_chat(user, SPAN_NOTICE("There is no gunshot residue on \the [O]."))
+			return TRUE
 		var/obj/C = H.get_covering_equipped_item_by_zone(zone)
-		if(C)
-			afterattack(C, user, 1) //Lazy but this would work
-			return
+		if (C)
+			use_after(C, user) //Lazy but this would work
+			return TRUE
 		user.visible_message("[user] swabs [H]'s [O.name] for a sample.")
 		sample_type = "gunshot_residue"
 		gunshot_residue_sample = O.gunshot_residue.Copy()
-
-	if(sample_type)
+	if (sample_type)
 		set_used(sample_type, H)
-		return
-	return 1
+	return TRUE
 
-/obj/item/forensics/swab/afterattack(atom/A, mob/user, proximity)
-
-	if(!proximity || istype(A, /obj/machinery/dnaforensics))
-		return
-
+/obj/item/forensics/swab/use_after(atom/A, mob/living/user, click_parameters)
 	if(is_used())
 		to_chat(user, SPAN_WARNING("This swab has already been used."))
-		return
-
-	add_fingerprint(user)
+		return TRUE
 
 	var/list/choices = list()
 	if(A.blood_DNA)
@@ -108,20 +99,20 @@
 	var/choice
 	if(!length(choices))
 		to_chat(user, SPAN_WARNING("There is no evidence on \the [A]."))
-		return
+		return TRUE
 	else if(length(choices) == 1)
 		choice = choices[1]
 	else
 		choice = input("What kind of evidence are you looking for?","Evidence Collection") as null|anything in choices
 
 	if(!choice)
-		return
+		return TRUE
 
 	var/sample_type
 	if(choice == "Blood")
 		if(!A.blood_DNA || !length(A.blood_DNA))
 			to_chat(user, SPAN_WARNING("There is no blood on \the [A]."))
-			return
+			return TRUE
 		dna = A.blood_DNA.Copy()
 		sample_type = "blood"
 
@@ -129,7 +120,7 @@
 		var/obj/item/clothing/B = A
 		if(!istype(B) || !B.gunshot_residue)
 			to_chat(user, SPAN_WARNING("There is no residue on \the [A]."))
-			return
+			return TRUE
 		gunshot_residue_sample = B.gunshot_residue.Copy()
 		sample_type = "residue"
 
@@ -137,13 +128,14 @@
 		var/obj/item/I = A
 		if(!istype(I) || !I.trace_DNA)
 			to_chat(user, SPAN_WARNING("There is no non-blood DNA on \the [A]."))
-			return
+			return TRUE
 		trace_dna = I.trace_DNA.Copy()
 		sample_type = "trace DNA"
 
 	if(sample_type)
 		user.visible_message("\The [user] swabs \the [A] for a sample.", "You swab \the [A] for a sample.")
 		set_used(sample_type, A)
+		return TRUE
 
 /obj/item/forensics/swab/proc/set_used(sample_str, atom/source)
 	SetName("[initial(name)] ([sample_str] - [source])")
